Do not use Unsafe type
As pointed out by others, this means the method table for public objects contain
a mix of public and private methods. That's bad. Instead, use things with
private names to do our dirty unsafe business.

This is an API change, so I'll leave it for a bit before merging, in case other
breaking stuff comes up.